% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexTest.R
\name{regexTest}
\alias{regexTest}
\title{A Gadget for testing gsub regular expressions.}
\usage{
regexTest(pattern = "night",
  x = "We can turn day into night with this Gadget", replace = "day")
}
\arguments{
\item{pattern}{The regular expression to match.}

\item{x}{The character string where matches are sought.}

\item{replacement}{The replacement text to substitute for the 
matching pattern.}
}
\description{
Supply a regex pattern, a string where matches
are sought, and a replacement. This gadget allows
regex testing. When the correct regex is found,
clicking "Done" will return the current regex.
}
\examples{

regexTest(pattern = 'night', replace = "day", x = "We can turn day into night with this Gadget")

}

